v1.3 feature: Error completed, filter supported

+++ src/main/java/org/cs209a/stackoverflowapp/entity/dto/AnswerQualityDTO.java
@@ -39,8 +39,8 @@ public class AnswerQualityDTO {
 
     @Data
     public static class GroupStats {
-        private int rangeStart;
-        private int rangeEnd;
+        private float rangeStart;
+        private float rangeEnd;
         private double avgQualityScore;
         private int count;
         private double acceptedRate;
